Web: types for plugins loader, removed circular dependency
